requirements.txt
pymysql==1.1.0
PyYAML==6.0.1
pandas==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
